+++ src/api/viewsets/project/service.py
@@ -1,6 +1,7 @@
+import os
 from typing import Dict
 
-from api.viewsets import Project
+from api.viewsets.project.model import Project
 from api.viewsets.project.serializer import ProjectSerializer
 
 
@@ -14,5 +15,15 @@ class ProjectService:
         return project_serializer.data
 
     @staticmethod
+    def get_by_id(project_id: int) -> Project:
+        return Project.objects.get(pk=project_id)
+
+    @staticmethod
     def delete(project_uuid: str):
         return Project.objects.filter(pk=project_uuid).delete()[0]
+
+    @staticmethod
+    def get_storage(storage_name: str) -> str:
+        if os.getenv('STORAGE_TYPE') == 'LOCAL':
+            return f"{os.getenv('STORAGE_PATH')}/{storage_name}"
+        return storage_name