+++ src/core/providers/file_system.py
@@ -1,0 +1,8 @@
+from core.managers.file import LocalFileManager
+
+
+class FileManagerProvider:
+
+    def __init__(self, storage):
+        if storage == 'LOCAL':
+            self.provider = LocalFileManager()