+++ src/tests/api/test_dataset.py
@@ -2,7 +2,8 @@ import pytest
 from model_bakery import baker
 from rest_framework.test import APIClient
 
-from api.viewsets import Dataset, Project
+from api.viewsets.dataset.model import Dataset
+from api.viewsets.project.model import Project
 
 
 @pytest.mark.django_db