+++ src/tests/fixtures/project.py
@@ -1,12 +1,12 @@
 import pytest
 from model_bakery import baker
 
-from api.viewsets import Project
+from api.viewsets.project.model import Project
 
 
 @pytest.fixture()
-def project() -> Project:
-    return baker.make(Project)
+def project(tmp_path) -> Project:
+    return baker.make(Project, storage_in=tmp_path)