+++ CHANGELOG.md
@@ -2,7 +2,17 @@
 
 All notable changes will be documented in this file
 
-## [0.0.8](https://github.com/pablobascunana/youml-manager/compare/90712cb...develop) (18-10-2022)
+## [0.0.9](https://github.com/pablobascunana/youml-manager/compare/2ed9f97...develop) (23-10-2022)
+
+### Added
+* [ML-75](https://youml.atlassian.net/browse/ML-75) Label model
+* [ML-77](https://youml.atlassian.net/browse/ML-77) Image model
+* [ML-78](https://youml.atlassian.net/browse/ML-78) ImageLabels model
+* [ML-79](https://youml.atlassian.net/browse/ML-79) Create image endpoint
+* [ML-92](https://youml.atlassian.net/browse/ML-92) Add file manager provider and local filesystem
+* [ML-76](https://youml.atlassian.net/browse/ML-76) Add storage_in field in project model and save it create project endpoint
+
+## [0.0.8](https://github.com/pablobascunana/youml-manager/compare/90712cb...2ed9f97) (18-10-2022)
 
 ### Added
 * [ML-68](https://youml.atlassian.net/browse/ML-68) Delete project